+++ zone/lua_mob.h
@@ -347,9 +347,9 @@ public:
 	void ChangeDrakkinHeritage(int in);
 	void ChangeDrakkinTattoo(int in);
 	void ChangeDrakkinDetails(int in);
-	void CameraEffect(uint32 duration, uint32 intensity);
-	void CameraEffect(uint32 duration, uint32 intensity, Lua_Client c);
-	void CameraEffect(uint32 duration, uint32 intensity, Lua_Client c, bool global);
+	void CameraEffect(uint32 duration, float intensity);
+	void CameraEffect(uint32 duration, float intensity, Lua_Client c);
+	void CameraEffect(uint32 duration, float intensity, Lua_Client c, bool global);
 	void SendSpellEffect(uint32 effect_id, uint32 duration, uint32 finish_delay, bool zone_wide,
 		uint32 unk020);
 	void SendSpellEffect(uint32 effect_id, uint32 duration, uint32 finish_delay, bool zone_wide,
